skill parser: split into testable CharacterSkills collaborators (#427)
* parser: fix description truncation + nested markup + default targets

split_top_level now treats [[…]] as nesting, so descriptions are no longer
truncated at pipes inside wikilinks (markup leakage 1.8%→0.02%).
display_description unwinds nested templates and strips wikilinks/orphan braces.
Default effect target by type when a clause has no subject (grant→caster,
inflict→one_foe, field→field) — nil targets 12.4%→0.02%.

* parser: golden-snapshot characterization spec

Locks the full observable parse output (slots, versions, effects with resolved
status names, links, EN+JP text) for the 5 sample characters. Regression net
for the upcoming #5 builder split — regenerate with REGENERATE=1.

* skill parser: extract CharacterSkills::FieldParser

Pure stateless wiki-text helpers (markup cleaning, display descriptions,
split_top_level, cooldown/level/duration field parsing) move out of the
parser, which now delegates. Align Metrics/ModuleLength with ClassLength (200).

* skill parser: extract CharacterSkills::EffectParser

Description -> ordered effect rows (clause-based targets, status matching,
damage/heal/dispel) moves out, owning the unmatched-status accumulator and the
TT/CONTEXT/DEFAULT_TARGETS/STACKING constants. Parser injects one instance.

* skill parser: extract CharacterSkills::JpLocalizer

JP-alignment (apply/group/localize/set_jp + STATE_VARIANT_ROLES) moves into its
own class that mutates the slot graph in place; the parser calls it in #parse.

* skill parser: extract CharacterSkills::Persister

The destroy-then-recreate graph persistence (slots/versions/effects/links +
permitted_attrs column filtering) moves into its own class.

* skill parser: extract CharacterSkills::Reporter

report_for + cross_validate_statuses + status_ailment_id move into a Reporter
(new(data:, status_lookup:)) that takes the graph and unmatched set as args.

* skill parser: extract CharacterSkills::Builder; slim facade

The heuristic core (slot/version building + role/trigger/progression inference)
moves into Builder, which owns @links/@version_keys and returns a graph with
links pre-filtered to built versions. CharacterSkillParser is now a thin façade
(~70 lines) wiring Builder/JpLocalizer/Reporter/Persister. Golden snapshot
unchanged — behavior preserved.

* skill parser: extract CharacterSkills::InferenceRules

Pure text->classification heuristics (variant role, trigger type/value,
uncap/transcendence progression, flag predicates) move out of the Builder into
a stateless module; the Builder delegates. Golden snapshot unchanged.
