+++ .rubocop.yml
@@ -38,6 +38,11 @@ Metrics/ClassLength:
   Exclude:
     - 'spec/**/*'
 
+Metrics/ModuleLength:
+  Max: 200
+  Exclude:
+    - 'spec/**/*'
+
 Metrics/AbcSize:
   Max: 40
   Exclude:
